ci: Fix Cypress cache location in forward merge script

+++ utils/forward-merge.js
@@ -187,7 +187,13 @@ async function main() {
     await fs.writeFile(file, newContents, 'utf8');
   }
 
-  await spawn(`yarn install --production=false`);
+  if (alreadyMerging) {
+    // already merging is a local dev
+    await spawn(`yarn install --production=false`);
+  } else {
+    // CI needs to set the Cypress cache folder
+    await spawn(`CYPRESS_CACHE_FOLDER=.cache/cypress yarn install --production=false`);
+  }
 
   if (hasConflicts) {
     if (hasUnresolvedConflicts) {
